Moved make df and make plot functions to utils folder.
